chore: adding minibatched rl task abstraction
